requirements.txt
APScheduler==3.9.1
discord_webhook==0.15.0
speedtest-cli==2.1.3
